Make the deliberate narrowings in three files explicit
-Wconversion's narrowing half reports 46 unique sites across
fdisk_screen.c, fdisk_memory.c and freeze_audiomix.c.  Every one seen so
far is already provably in range -- a mask, a shift, or a loop bound
does the proving and the compiler simply does not track it.

fdisk_screen.c and fdisk_memory.c are now clean; freeze_audiomix.c goes
from 37 to 17.  All seven targets are byte-identical, so the casts cost
nothing.

The cast is chosen over a pragma because it asserts something a reader
can check against the mask beside it.  It is not free of consequence:
an explicit cast also hides that site from tidy-arithmetic permanently,
which is why it belongs only where the range is provable locally.

+++ src/fdisk_memory.c
@@ -57,9 +57,9 @@ __attribute__((noinline)) void lcopy(
     }
     dmalist.option_0b = 0x0b;
     dmalist.option_80 = 0x80;
-    dmalist.source_mb = source_address >> 20;
+    dmalist.source_mb = (uint8_t)(source_address >> 20);
     dmalist.option_81 = 0x81;
-    dmalist.dest_mb = (destination_address >> 20);
+    dmalist.dest_mb = (uint8_t)(destination_address >> 20);
     dmalist.end_of_options = 0x00;
 
     dmalist.command = 0x00; // copy
@@ -88,7 +88,7 @@ __attribute__((noinline)) void lfill(
     dmalist.option_80 = 0x80;
     dmalist.source_mb = 0x00;
     dmalist.option_81 = 0x81;
-    dmalist.dest_mb = destination_address >> 20;
+    dmalist.dest_mb = (uint8_t)(destination_address >> 20);
     dmalist.end_of_options = 0x00;
 
     dmalist.command = 0x03; // fill
